final commit (for now)
app does everything it's supposed to, by the specs at least. there are 2 bugs, but that is due to the names of the airlines so sometimes the logos don't show as well as some flight information, but that is because there is no information about those flights. as for the logos, i have found a solution but that would require inserting additional stuff like node.js but for some reason php storm was determined not to allow me to use required function so i couldn't do that part.